Improve X and Y input calculations for joystick corners.
When the joystick is moved into a corner, the X and Y vectors should not
be added linearly. Doing so results in a disproportionate amount of force
being applied to in-game movement at 45-degree angles.

This leads to problems in Ocarina of Time, for example.
By slightly moving the joystick horizontally or vertically, Link will
slowly and carefully tiptoe into the indicated direction. But tiptoeing
along a path at a 45-degree angle is impossible because Link will start
jogging as soon as the joystick is moved.
(Without this patch, the problem is especially visible if X and Y inputs
are capped at a low value such as 20 or 40.)

This behaviour makes it impossible to get at the Megaton Hammer in the
Fire Temple. A narrow winding staircase has to be climbed quickly within
about 15 seconds, but Link will all too easily fall off the edge if the
game is played with the input-sdl plugin.

This patch smoothes X and Y inputs in corners such that the directional
forces applied by the joystick form a circle, rather than a square.
This matches the N64 hardware behaviour more closely.
The Megaton Hammer is now within reach!

Patch by myself and Neels Hofmeyr.