Fix Xbox 360 Controller for Windows Mappings
I'm almost scared to make this pull request. This has got to be one of the most tested autoconfigs listed. But the mappings were wayyy off in three different Windows 10 machines. The only thing I've set in my mappings that could be up for debate is that I've set C Left and C Up to X and Y, respectively. I chose this because they sit in ergonomically similar positions (And navi really wants you to use C Up).